+++ src/components/action-popover/action-popover.stories.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
 import {
-  ActionPopover, ActionPopoverDivider, ActionPopoverItem, ActionPopoverMenu
+  ActionPopover, ActionPopoverDivider, ActionPopoverItem, ActionPopoverMenu, ActionPopoverMenuButton
 } from '.';
 import { dlsThemeSelector, classicThemeSelector } from '../../../.storybook/theme-selectors';
 import {
   Table, TableRow, TableCell, TableHeader
 } from '../table';
-import Button from '../button';
 
 const submenu = (
   <ActionPopoverMenu>
@@ -214,18 +213,17 @@ export const StylesOverriden = () => (
         <TableCell>
           <ActionPopover
             rightAlignMenu
-            renderButton={ ({ styleOverride, tabIndex, ...rest }) => (
-              <Button
+            renderButton={ ({ tabIndex, 'data-element': dataElement }) => (
+              <ActionPopoverMenuButton
                 buttonType='tertiary'
                 iconType='dropdown'
                 iconPosition='after'
                 size='small'
                 tabIndex={ tabIndex }
-                styleOverride={ styleOverride }
-                date-element={ rest['data-element'] }
+                data-element={ dataElement }
               >
                 More
-              </Button>
+              </ActionPopoverMenuButton>
             ) }
           >
             <ActionPopoverItem onClick={ action('auto entry') }>Auto Entry</ActionPopoverItem>
@@ -243,18 +241,17 @@ export const StylesOverriden = () => (
         <TableCell>
           <ActionPopover
             rightAlignMenu
-            renderButton={ ({ styleOverride, tabIndex, ...rest }) => (
-              <Button
+            renderButton={ ({ tabIndex, 'data-element': dataElement }) => (
+              <ActionPopoverMenuButton
                 buttonType='tertiary'
                 iconType='dropdown'
                 iconPosition='after'
                 size='small'
                 tabIndex={ tabIndex }
-                styleOverride={ styleOverride }
-                date-element={ rest['data-element'] }
+                data-element={ dataElement }
               >
                 More
-              </Button>
+              </ActionPopoverMenuButton>
             ) }
           >
             <ActionPopoverItem onClick={ action('auto entry') }>Auto Entry</ActionPopoverItem>